#93 Move password check code so it can be used in reset password.
