+++ api/src/membership/member_auth.py
@@ -6,16 +6,30 @@ from service.db import db_session
 from service.error import Unauthorized
 
 
-def get_member_permissions(member_id=None):
-    """ Return query to get all (permission_id, permission) for a memeber, used from core. """
-    return (
-        db_session
-            .query(Permission.permission_id, Permission.permission)
-            .distinct()
-            .join(Group, Permission.groups)
-            .join(Member, Group.members)
-            .filter_by(member_id=member_id)
-    )
+ALPHA_LOWER_SET = set("abcdefghijklmnopqrstuvwxyzåäö")
+ALPHA_UPPER_SET = set("ABCDEFGHIJKLMNOPQRSTUVWXYZÅÄÖ")
+DIGIT_SET = set("0123456789")
+ALPHA_DIGIT_SET = ALPHA_LOWER_SET | ALPHA_UPPER_SET | DIGIT_SET
+
+FORBIDDEN_SUB_SEQUENCES = [
+    ("abcdefghijklmnopqrstuvwxyzåäö", 4),
+    ("abcdefghijklmnopqrstuvwxyzåäö"[::-1], 4),
+    ("qwertyuiopå", 4),
+    ("qwertyuiopå"[::-1], 4),
+    ("asdfghjklöä", 4),
+    ("asdfghjklöä"[::-1], 4),
+    ("zxcvbnm,.", 4),
+    ("zxcvbnm,."[::-1], 4),
+    ("01234567890", 3),
+    ("01234567890"[::-1], 3),
+]
+
+
+def contains_sub_sequence(value, sequence, length):
+    for i in range(0, len(sequence) - length + 1):
+        if value.find(sequence[i:i + length]) != -1:
+            return True
+    return False
 
 
 def verify_password(password, password_hash):
@@ -26,6 +40,40 @@ def verify_password(password, password_hash):
 
 def hash_password(password):
     return bcrypt.hashpw(password=password.encode(), salt=bcrypt.gensalt()).decode()
+
+
+def check_and_hash_password(unhashed_password):
+    if unhashed_password is None:
+        return None
+        
+    contained_chars_set = set(unhashed_password)
+
+    if len(unhashed_password) < 8 or len(contained_chars_set) < 6:
+        raise ValueError("password must be at least 8 characters long with at least 6 unique characters")
+
+    # Test for forbidden sequences
+    forbidden_sequences = {
+        "password", "q1w2e3r4t5", "maker", "space"
+    }
+    if (
+            any([contains_sub_sequence(unhashed_password, s, l) for s, l in FORBIDDEN_SUB_SEQUENCES]) or
+            any([unhashed_password.find(s) != -1 for s in forbidden_sequences])
+    ):
+        raise ValueError("password contains a forbidden sequence of characters, try something less common")
+
+    return hash_password(unhashed_password)
+
+
+def get_member_permissions(member_id=None):
+    """ Return query to get all (permission_id, permission) for a memeber, used from core. """
+    return (
+        db_session
+            .query(Permission.permission_id, Permission.permission)
+            .distinct()
+            .join(Group, Permission.groups)
+            .join(Member, Group.members)
+            .filter_by(member_id=member_id)
+    )
 
 
 def authenticate(username=None, password=None):
